Update documentation and fix code quality issues
- Updated README.md with correct test count (41 tests: 28 protocol validation + 13 block integration)
- Updated examples/README.md to document MMDVM protocol blocks (POCSAG, D-STAR, YSF, P25)
- Fixed flake8 issues: removed unused imports, fixed whitespace
- Applied black formatting to Python files
- All 41 tests passing, code quality checks passing

+++ python/qradiolink/bindings/demod_freedv_python.cc
@@ -30,7 +30,7 @@ void bind_demod_freedv(py::module& m)
              py::arg("carrier_freq") = 1700,
              py::arg("filter_width") = 2000,
              py::arg("low_cutoff") = 200,
-             py::arg("mode") = gr::vocoder::freedv_api::MODE_1600,
+             py::arg("mode") = static_cast<int>(gr::vocoder::freedv_api::MODE_1600),
              py::arg("sb") = 0,
              "Make a FreeDV demodulator block")
